Much simplified.
The indexing of articles is now done entirely in a tree method, which
means the articles don't have to know about it and the route function
doesn't need to bother with it.

Still generating all variables for all articles to begin with, which is
of course massively inefficient. Haven't quite found a way to get around
that yet, as articles need to be able to access the variables from other
articles.